+++ jest.config.native.js
@@ -40,5 +40,6 @@ module.exports = {
         '<rootDir>/../../suite-native/analytics/src/jest.setup.ts',
         '<rootDir>/../../suite-native/module-trading/src/jest.setup.tsx',
         '<rootDir>/../../suite-native/module-connect-popup/src/jest.setup.ts',
+        '<rootDir>/../../suite-native/module-device-onboarding/src/jest.setup.ts',
     ],
 };
