+++ suite-native/module-accounts-management/src/components/AccountSettingsShowXpubButton.tsx
@@ -1,54 +1,19 @@
 import { useCallback, useState } from 'react';
 import { useSelector } from 'react-redux';
 
-import { TrezorDevice } from '@suite-common/suite-types';
 import {
     AccountsRootState,
     selectAccountByKey,
     selectIsDeviceBackupRequired,
     selectSelectedDevice,
+    showXpubOnDevice,
 } from '@suite-common/wallet-core';
-import { Account } from '@suite-common/wallet-types';
-import { getDerivationType, isAddressBasedNetwork } from '@suite-common/wallet-utils';
+import { isAddressBasedNetwork } from '@suite-common/wallet-utils';
 import { Button, useBottomSheetModal } from '@suite-native/atoms';
 import { Translation } from '@suite-native/intl';
 import { WalletBackupNotSetWarningBottomSheet } from '@suite-native/module-device-onboarding';
 import { XpubQRCodeBottomSheet } from '@suite-native/qr-code';
-import TrezorConnect, { Success, Unsuccessful } from '@trezor/connect';
 import { convertTaprootXpub } from '@trezor/utils';
-
-export const showXpubOnDevice = async (device: TrezorDevice, account: Account) => {
-    if (!device || !account) return;
-
-    const params = {
-        device,
-        path: account.path,
-        useEmptyPassphrase: device.useEmptyPassphrase,
-        showOnTrezor: true,
-        derivationType: getDerivationType(account.accountType),
-        coin: account.symbol,
-    };
-
-    let response: Success<unknown> | Unsuccessful;
-    switch (account.networkType) {
-        case 'bitcoin':
-            response = await TrezorConnect.getPublicKey(params);
-            break;
-        case 'cardano':
-            response = await TrezorConnect.cardanoGetPublicKey(params);
-            break;
-        case 'solana':
-            response = await TrezorConnect.solanaGetPublicKey(params);
-            break;
-        default:
-            response = {
-                success: false,
-                payload: { error: 'Method for getPublicKey not defined', code: undefined },
-            };
-    }
-
-    return response;
-};
 
 export const AccountSettingsShowXpubButton = ({ accountKey }: { accountKey: string }) => {
     const account = useSelector((state: AccountsRootState) =>
