+++ suite-native/module-device-onboarding/src/jest.setup.ts
@@ -1,0 +1,3 @@
+jest.mock('@suite-native/module-device-onboarding', () => ({
+    WalletBackupNotSetWarningBottomSheet: () => null,
+}));
